Fix: fixed the component organizations and header completed
