+++ app/components/organisms/headerComponents/components/header.style.ts
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import { sizes } from "@/styles/sizes";
-import HamburgerButton from "@/app/components/organisms/navbar/mobileNavbar/hamburgerButton/hamburgerButton";
+import HamburgerButton from "@/app/components/organisms/headerComponents/curtainMenu/navbar/mobileNavbar/hamburgerButton/hamburgerButton";
 
 export const HeaderWrapper = styled.div`
   display: flex;
