+++ app/components/organisms/headerComponents/curtainMenu/navbar/mobileNavbar/component/components/components/subList.style.ts
@@ -6,10 +6,11 @@ export const SublistWrapper = styled.ul`
   flex-direction: column;
   padding: 16px 0 8px 8px;
   gap: ${sizes.xs}px;
+  background-color: ${({ theme }) => theme.colors.grey[50]};
   @media (min-width: 1024px) {
     gap: ${sizes.sm}px;
     padding: 16px 0 16px 16px;
     width: ${sizes["5xl"]}px;
-    box-shadow: 5px 11px 8px 2px rgba(0, 0, 0, 0.1);
+    box-shadow: 5px 5px 8px 1px rgba(0, 0, 0, 0.1);
   }
 `;
