+++ app/components/organisms/headerComponents/curtainMenu/navbar/mobileNavbar/hamburgerButton/component/hamburgerButton.style.ts
