+++ app/components/organisms/headerComponents/curtainMenu/navbar/navbar.style.ts
@@ -6,8 +6,10 @@ export const NavbarWrapper = styled.nav`
   flex-direction: column;
   gap: ${sizes.sm}px;
   @media (min-width: 376px) {
+    padding-top: ${sizes.lg}px;
   }
   @media (min-width: 1024px) {
     flex-direction: row;
+    padding-top: ${sizes.xs}px;
   }
 `;
