--- app/components/organisms/navbar/mobileNavbar/component/mobileNav.tsx
@@ -1,11 +1,0 @@
-import { menuData } from "./data";
-import { NavbarWrapper } from "./mobileNav.style";
-import NavList from "./navList";
-
-export default function MobileNavbar() {
-  return (
-    <NavbarWrapper>
-      <NavList />
-    </NavbarWrapper>
-  );
-}
